+++ tests/e2e/utils/pageObjects/sar/sarDashboard.page.ts
@@ -1,7 +1,7 @@
 import { Locator, Page } from "@playwright/test";
 import BasePage from "../base.page";
-import { firstPeriod, stateAbbreviation, stateName } from "../../helpers";
-import { currentYear } from "../../../seeds/helpers";
+import { firstPeriod, stateAbbreviation, stateName } from "../../../helpers";
+import { currentYear } from "../../../../seeds/helpers";
 
 export class SARDashboardPage extends BasePage {
   public path = "/sar";
